requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
Pillow>=10.0.0
uwsgi
django-jazzmin
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0